fix: fix ci
